feat: add myServer
